+++ src/js/visual/configurator-csr.js
@@ -65,6 +65,7 @@ ripe.ConfiguratorCsr.prototype.init = function() {
     this.pixelRatio =
         this.options.pixelRatio || (typeof window !== "undefined" && window.devicePixelRatio) || 2;
     this.sensitivity = this.options.sensitivity || 40;
+    this.duration = this.options.duration || 500;
     const rendererOpts = this.options.rendererOptions || {};
     this.rendererOptions = {
         outputEncoding:
@@ -92,20 +93,26 @@ ripe.ConfiguratorCsr.prototype.init = function() {
         far: cameraOpts.far !== undefined ? cameraOpts.far : 10000,
         posX: cameraOpts.posX !== undefined ? cameraOpts.posX : 0,
         posY: cameraOpts.posY !== undefined ? cameraOpts.posY : 0,
-        posZ: cameraOpts.posZ !== undefined ? cameraOpts.posZ : 50
+        posZ: cameraOpts.posZ !== undefined ? cameraOpts.posZ : 6
     };
 
     // general state variables
     this.loading = true;
+    this.noDrag = false;
     this.currentSize = 0;
     this.currentWidth = 0;
     this.currentHeight = 0;
+    this.clock = null;
+    this.animations = [];
+    this.isChangeFrameAnimationRunning = false;
+    this._pendingOps = [];
 
     // CSR variables
     this.renderer = null;
     this.camera = null;
     this.scene = null;
     this.environmentTexture = null;
+    this.modelGroup = null;
     this.mesh = null;
 
     // handlers variables
@@ -118,8 +125,9 @@ ripe.ConfiguratorCsr.prototype.init = function() {
     // creates the necessary DOM elements and runs the
     // CSR initializer
     this._initLayout();
-    this._initCsr().then(() => {
+    this._initCsr().then(async () => {
         this.loading = false;
+        await this.flushPending(true);
     });
 };
 
@@ -156,6 +164,7 @@ ripe.ConfiguratorCsr.prototype.updateOptions = async function(options, update = 
     this.size = options.size === undefined ? this.size : options.size;
     this.pixelRatio = options.pixelRatio === undefined ? this.pixelRatio : options.pixelRatio;
     this.sensitivity = options.sensitivity === undefined ? this.sensitivity : options.sensitivity;
+    this.duration = options.duration === undefined ? this.duration : options.duration;
     const rendererOpts = options.rendererOptions || {};
     this.rendererOptions = { ...this.rendererOptions, ...rendererOpts };
     this.useDracoLoader =
@@ -244,6 +253,173 @@ ripe.ConfiguratorCsr.prototype.resize = async function(size, width, height) {
             force: true
         }
     );
+};
+
+/**
+ * Executes pending operations that were not performed so as
+ * to not conflict with the tasks already being executed.
+ *
+ * The main reason for collision is the loading operation
+ * being executed (long duration operation).
+ *
+ * @param {Boolean} tail If only the last pending operation should
+ * be flushed, meaning that the others are discarded.
+ */
+ripe.ConfiguratorCsr.prototype.flushPending = async function(tail = false) {
+    const pending =
+        tail && this._pendingOps.length > 0
+            ? [this._pendingOps[this._pendingOps.length - 1]]
+            : this._pendingOps;
+    this._pendingOps = [];
+    while (pending.length > 0) {
+        const { operation, args } = pending.shift();
+        switch (operation) {
+            case "changeFrame":
+                await this.changeFrame(...args);
+                break;
+            default:
+                break;
+        }
+    }
+};
+
+/**
+ * Rotates the model to match the PRC frame.
+ *
+ * @param {Object} frame The new PRC frame to display using the extended and canonical
+ * format for the frame description (eg: side-3).
+ * @param {Object} options Set of optional parameters to adjust the change frame, such as:
+ * - 'duration' - The duration of the animation in milliseconds (defaults to 'null').
+ * - 'stepDuration' - If defined the total duration of the animation is
+ * calculated using the amount of steps times the number of steps, instead of
+ * using the 'duration' field (defaults to 'null').
+ * - 'revolutionDuration' - If defined the step duration is calculated by dividing
+ * the revolution duration by the number of frames in the view (defaults to 'null').
+ * - 'preventDrag' - If drag actions during an animated change of frames should be
+ * ignored (defaults to 'true').
+ * - 'safe' - If requested then the operation is only performed in case the configurator
+ * is not in the an equivalent state (default to 'true').
+ */
+ripe.ConfiguratorCsr.prototype.changeFrame = async function(frame, options = {}) {
+    // in case the scene group is not loaded then it's not possible to change the frame
+    if (!this.modelGroup) throw new Error("Model group not loaded");
+
+    // parses the requested frame value according to the pre-defined
+    // standard (eg: side-3) and then unpacks it as view and position
+    const _frame = ripe.parseFrameKey(frame);
+    const nextView = _frame[0];
+    const nextPosition = parseInt(_frame[1]);
+
+    // in case the next position value was not properly parsed (probably undefined)
+    // then it's not possible to change frame (throws exception)
+    if (isNaN(nextPosition)) {
+        throw new RangeError("Frame position is not defined");
+    }
+
+    // unpacks the other options to the frame change defaulting their values
+    // in case undefined values are found
+    let duration = options.duration !== undefined ? options.duration : null;
+    let stepDuration = options.stepDuration !== undefined ? options.stepDuration : null;
+    const revolutionDuration =
+        options.revolutionDuration !== undefined ? options.revolutionDuration : null;
+    const preventDrag = options.preventDrag !== undefined ? options.preventDrag : true;
+    const safe = options.safe !== undefined ? options.safe : true;
+
+    // retrieves the model frame object
+    const frames = await this.owner.getFrames();
+
+    // tries to retrieve the amount of frames for the target view and
+    // validates that the target view exists and that the target position
+    // (frame) does not overflow the amount of frames in for the view
+    const viewFramesNum = frames[nextView];
+    if (!viewFramesNum || nextPosition >= viewFramesNum) {
+        throw new RangeError("Frame " + frame + " is not supported");
+    }
+
+    // in case the safe mode is enabled and the current configuration is
+    // still under the loading situation the change frame operation is
+    // saved and will be executed after the loading is finished
+    if (safe && this.loading) {
+        this._pendingOps = [{ operation: "changeFrame", args: [frame, options] }];
+        return;
+    }
+
+    // normalizes the model group rotation values
+    ripe.CsrUtils.normalizeRotations(this.modelGroup);
+
+    // calculates step duration based on revolutionDuration defaulting to the stepDuration
+    // if no revolutionDuration is specified
+    stepDuration =
+        revolutionDuration !== null ? parseInt(revolutionDuration / viewFramesNum) : stepDuration;
+
+    // bases the duration on the stepDuration if stepDuration is specified
+    if (stepDuration) {
+        // default the step count to 1
+        let stepCount = 1;
+
+        // step count calculation logic for "side" view
+        if (nextView === "side") {
+            // calculates the PRC compatible step count for the change frame animation
+            const radPerSide = (2 * Math.PI) / viewFramesNum;
+            const rotYStart = ripe.CsrUtils.toPrecision(this.modelGroup.rotation.y);
+            const rotYEnd = ripe.CsrUtils.toPrecision(nextPosition * radPerSide);
+            const rotYQty = rotYEnd - rotYStart;
+
+            // gets the number of PRC compatible steps
+            stepCount = Math.abs(rotYQty / radPerSide);
+        }
+
+        // rounds up the step count to it's respective a whole number
+        const stepCountRounded = Math.ceil(stepCount);
+
+        // sets duration based on stepDuration
+        duration = stepDuration * stepCountRounded;
+    }
+
+    // ensures duration default if no duration is set
+    duration = duration !== null ? duration : this.duration;
+
+    // duration value compatible with CSR animation defaulting
+    // to 0 if no duration was successfully set
+    duration = duration ? duration / 1000 : 0;
+
+    // creates a change frame animation
+    const animation = new ripe.CsrChangeFrameAnimation(
+        this.modelGroup,
+        duration,
+        nextView,
+        nextPosition,
+        viewFramesNum
+    );
+
+    // sets the animation to clean the state when the animation finishes
+    animation.onFinished = () => {
+        this.isChangeFrameAnimationRunning = false;
+        this.noDrag = false;
+    };
+
+    // checks for change frame animations that are already running
+    const index = this.animations.findIndex(
+        animation => animation.type === "CsrChangeFrameAnimation"
+    );
+    const runningAnimation = this.animations[index];
+    if (runningAnimation) {
+        // ignore this call as the running animation is the same
+        if (runningAnimation.signature === animation.signature) {
+            this.isChangeFrameAnimationRunning = false;
+            this.noDrag = false;
+            return;
+        }
+
+        // removes the running animation as the new animation is different
+        this.animations.splice(index, 1);
+    }
+
+    // sets the needed state variables and adds the new animation so
+    // it can be run
+    this.isChangeFrameAnimationRunning = true;
+    this.noDrag = preventDrag;
+    this.animations.push(animation);
 };
 
 /**
@@ -337,15 +513,23 @@ ripe.ConfiguratorCsr.prototype._loadEnvironment = async function(path) {
  * @private
  */
 ripe.ConfiguratorCsr.prototype._loadScene = async function() {
+    // inits the scene clock
+    this.clock = new window.THREE.Clock();
+
     // loads and sets scene environment
     this.environmentTexture = await this._loadEnvironment(this.sceneEnvironmentPath);
     this.environmentTexture.mapping = window.THREE.EquirectangularReflectionMapping;
     this.scene.environment = this.environmentTexture;
 
+    // inits the scene model group
+    this.modelGroup = new window.THREE.Group();
+
     // loads and sets the model mesh
     const meshPath = this.owner.getMeshUrl();
     this.mesh = await this._loadMesh(meshPath);
-    this.scene.add(this.mesh);
+    this.modelGroup.add(this.mesh);
+
+    this.scene.add(this.modelGroup);
 };
 
 /**
@@ -455,11 +639,14 @@ ripe.ConfiguratorCsr.prototype._deinitCsr = function() {
         this.environmentTexture = null;
     }
 
-    if (this.mesh) {
-        if (this.mesh.geometry) this.mesh.geometry.dispose();
-        if (this.mesh.material) this.mesh.material.dispose();
-        if (this.scene) this.scene.remove(this.mesh);
-        this.mesh = null;
+    if (this.modelGroup) {
+        if (this.mesh) {
+            if (this.mesh.geometry) this.mesh.geometry.dispose();
+            if (this.mesh.material) this.mesh.material.dispose();
+            this.modelGroup.remove(this.mesh);
+            this.mesh = null;
+        }
+        this.modelGroup = null;
     }
 
     if (this.scene) this.scene = null;
@@ -499,6 +686,28 @@ ripe.ConfiguratorCsr.prototype._resizeCsr = function(width, height) {
  * @private
  */
 ripe.ConfiguratorCsr.prototype._onAnimationLoop = function(self) {
+    if (self.loading) return;
+
+    // processes the animation loop tick delta
+    const delta = self.clock.getDelta();
+
+    if (!self.modelGroup) return;
+
+    // checks if there are animations to process
+    if (self.animations.length > 0) {
+        // ticks animations
+        for (let i = self.animations.length - 1; i >= 0; i--) {
+            const animation = self.animations[i];
+            animation.tick(delta);
+
+            if (animation.isFinished()) self.animations.splice(i, 1);
+        }
+
+        // normalizes the model group rotations
+        ripe.CsrUtils.normalizeRotations(self.modelGroup);
+    }
+
+    // renders a frame
     self._render();
 };
 
@@ -536,7 +745,8 @@ ripe.ConfiguratorCsr.prototype._onMouseLeave = function(self, event) {
  */
 ripe.ConfiguratorCsr.prototype._onMouseMove = function(self, event) {
     if (!self.isMouseDown) return;
-    if (!self.mesh) return;
+    if (!self.modelGroup) return;
+    if (self.noDrag) return;
 
     const mousePosX = event.pageX;
     const mousePosY = event.pageY;
@@ -549,10 +759,10 @@ ripe.ConfiguratorCsr.prototype._onMouseMove = function(self, event) {
     const sensitivity = self.sensitivity * 0.1;
 
     const dragValueX = (percentX - self.prevPercentX) * sensitivity;
-    self.mesh.rotation.y -= dragValueX;
+    self.modelGroup.rotation.y -= dragValueX;
 
     const dragValueY = (percentY - self.prevPercentY) * sensitivity;
-    self.mesh.rotation.x -= dragValueY;
+    self.modelGroup.rotation.x -= dragValueY;
 
     self.prevPercentX = percentX;
     self.prevPercentY = percentY;