+++ CHANGELOG.md
@@ -13,11 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add `variant` passing through `onConfig` and `onPart`
 * Add CSS `link` to `README.md`
 * Add `hasStrategy` method
+* Add `CsrAnimation`, `CsrChangeFrameAnimation` and `CsrUtils`
+* Add `changeFrame` and `flushPending` methods to CSR configurator
+* Add click binds to ripe-sdk-demo that allows to match CSR to the PRC frames
 
 ### Changed
 
 * Avoid warnings by not using `var` in ripe-sdk-demo .js files
 * Set renderer output encoding to `sRGBEncoding`
+* Set ripe-sdk-demo brand and model to `dummy` `cube`
+* Set CSR camera default values standard
+* Use `THREE.Group` to manipulate positions and rotations instead of directly changing mesh axis
+* Improve animation loop so it can tick animations
 
 ### Fixed